+++ feature-docs/.prompts.md
@@ -4,7 +4,7 @@ Create a pom.xml for a project running a Undertow http server, using Java versio
 maven.compiler.* properties.
 groupid should be net.stoerr.chatgpt, artifactid developersChatGPTtoolbenchPlugin.
 
-Create a java file net.stoerr.chatgpt.devtoolbench.DevToolBench that starts a http server and returns "Hello World!" to
+Create a java file net.stoerr.chatgpt.codevengine.DevToolBench that starts a http server and returns "Hello World!" to
 every get request on port 8080.
 Configure that as the main class for the maven generated jar.
 Use the shade plugin to create a jar that has all dependencies included.
