+++ pom.xml
@@ -33,7 +33,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-        <mainClass>net.stoerr.chatgpt.devtoolbench.DevToolBench</mainClass>
+        <mainClass>net.stoerr.chatgpt.codevengine.DevToolBench</mainClass>
     </properties>
 
     <dependencies>
