+++ src/main/java/net/stoerr/chatgpt/codevengine/ExecutionAbortedException.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 /**
  * Thrown to abort an execution, but only when the error has been duly reported with sendError to ChatGPT, so that the exception can be ignored.
