+++ src/main/java/net/stoerr/chatgpt/codevengine/ListFilesAction.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
