+++ src/main/java/net/stoerr/chatgpt/codevengine/RepeatedRequestChecker.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
