+++ src/main/java/net/stoerr/chatgpt/codevengine/UrlAction.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
