+++ src/test/java/net/stoerr/chatgpt/codevengine/DevToolBenchGeneralIT.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
