+++ src/test/java/net/stoerr/chatgpt/codevengine/GrepActionIT.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import static org.hamcrest.CoreMatchers.is;
