+++ src/test/java/net/stoerr/chatgpt/codevengine/TbUtilsTest.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
