+++ src/test/java/net/stoerr/chatgpt/codevengine/WriteFileActionIT.java
@@ -1,4 +1,4 @@
-package net.stoerr.chatgpt.devtoolbench;
+package net.stoerr.chatgpt.codevengine;
 
 import java.nio.file.Files;
 import java.nio.file.Paths;
